Move função de converter imagem para base64 para pasta utils
