+++ resources/js/pages/edit-profile/index.js
@@ -1,5 +1,4 @@
-import { imageToBase64 } from "./src/index.js";
-import { hiddenDropdown, toggleDropdownVisibilityHandler } from '../utils/src/index.js';
+import { hiddenDropdown, toggleDropdownVisibilityHandler, imageToBase64 } from '../utils/src/index.js';
 
 const photoInput = document.querySelector('input[name="photo"]');
