--- resources/js/pages/edit-profile/src/index.js
@@ -1,8 +1,0 @@
-const imageToBase64 = (image) => new Promise((resolve, reject) => {
-    const reader = new FileReader();
-    reader.readAsDataURL(image);
-    reader.onload = () => resolve(reader.result);
-    reader.onerror = error => reject(error);
-});
-
-export {imageToBase64}
